# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Number masking service",
    "description": "Proxy-number voice masking: two parties talk through a DID without learning each other's real number.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sessions/{session_id}/calls": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "Session Calls",
        "description": "Returns the call journal of one session.",
        "operationId": "session_calls_api_v1_sessions__session_id__calls_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calls": {
      "get": {
        "tags": [
          "calls"
        ],
        "summary": "List Calls",
        "description": "Returns a page of the journal across all sessions, newest first.",
        "operationId": "list_calls_api_v1_calls_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/numbers": {
      "get": {
        "tags": [
          "numbers"
        ],
        "summary": "Get Pool",
        "description": "Returns the pool: totals plus one entry per number.",
        "operationId": "get_pool_api_v1_numbers_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "numbers"
        ],
        "summary": "Add Number",
        "description": "Adds a number to the pool.",
        "operationId": "add_number_api_v1_numbers_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NumberCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/numbers/{e164}": {
      "patch": {
        "tags": [
          "numbers"
        ],
        "summary": "Update Number",
        "description": "Enables or disables a number.\n\nDisabling never breaks live sessions; it only removes the number from\nfuture allocations.",
        "operationId": "update_number_api_v1_numbers__e164__patch",
        "parameters": [
          {
            "name": "e164",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "E164"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NumberStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CallListOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CallOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "CallListOut",
        "description": "A page of the call journal."
      },
      "CallOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "proxy_number": {
            "type": "string",
            "title": "Proxy Number"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "answered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answered At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "hangup_cause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hangup Cause"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "direction",
          "proxy_number",
          "started_at",
          "answered_at",
          "ended_at",
          "duration_sec",
          "status",
          "hangup_cause"
        ],
        "title": "CallOut",
        "description": "One journal entry. Carries no real phone number."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NumberCreate": {
        "properties": {
          "e164": {
            "type": "string",
            "maxLength": 20,
            "title": "E164",
            "examples": [
              "+77011234567"
            ]
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "status": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Status",
            "default": "enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "e164"
        ],
        "title": "NumberCreate",
        "description": "Body of ``POST /numbers``."
      },
      "NumberOut": {
        "properties": {
          "e164": {
            "type": "string",
            "title": "E164"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "active_sessions": {
            "type": "integer",
            "title": "Active Sessions"
          },
          "released_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          },
          "in_cooldown": {
            "type": "boolean",
            "title": "In Cooldown"
          }
        },
        "type": "object",
        "required": [
          "e164",
          "status",
          "provider",
          "active_sessions",
          "released_at",
          "in_cooldown"
        ],
        "title": "NumberOut",
        "description": "One proxy number and its current occupancy."
      },
      "NumberStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "NumberStatusUpdate",
        "description": "Body of ``PATCH /numbers/{e164}``."
      },
...
      "PoolOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "enabled": {
            "type": "integer",
            "title": "Enabled"
          },
          "disabled": {
            "type": "integer",
            "title": "Disabled"
          },
          "in_use": {
            "type": "integer",
            "title": "In Use"
          },
          "in_cooldown": {
            "type": "integer",
            "title": "In Cooldown"
          },
          "free": {
            "type": "integer",
            "title": "Free"
          },
          "active_sessions": {
            "type": "integer",
            "title": "Active Sessions"
          },
          "numbers": {
            "items": {
              "$ref": "#/components/schemas/NumberOut"
            },
            "type": "array",
            "title": "Numbers"
          }
        },
        "type": "object",
        "required": [
          "total",
          "enabled",
          "disabled",
          "in_use",
          "in_cooldown",
          "free",
          "active_sessions",
          "numbers"
        ],
        "title": "PoolOut",
        "description": "The whole proxy number pool with aggregate counters."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}